Define a new backend interface, split over several roles
After lengthy discussions with DOLMEN, I decided to remove
the StoreFront/StoreBack complexity, and have the database
object simply interact with a single backend object.

The backend object will itself interact with a git wrapper
to provide one or more of the methods defined in the roles
(the only required being Git::Database::Role::Backend).
The more roles a backend supports, the more complete it
will be.

Git::Repository is not a requirement any more.