Prevent infinite loops by dying early
If none of digest, content and the eventual extra attribute is defined,
we will enter and infinite loop of deep recursion when trying to get any
attribute, as the builder will try to get another missing attribute,
and so on until it loops back to the attribute that was requested in
the first place.

The BUILD check will die when none of the attributes used by the various
builders is set.