Make Git::Database a factory for Git::Database::Backend::* objects
Git::Database was already vey thin, this makes it basically disappear.

Instead of delegating all the methods to backends that may not support them,
and then overriding can() in a weird way, we just make Git::Database->new
return the backend objects directly.

Checking if a method is supported is just a straightforward call to
the usual can() method.