Drop the 'backend' parameter from Git::Database->new
There's no point to having a 'backend' parameter when all
it does is returm the argument itself.

Note: I resisted the temptation to add 'git_dir' and 'work_tree',
along with a new meaning for 'backend' (the backend class name).

The idea was to delegate the creation of the 'store' object by
passing the 'git_dir' and 'work_tree' attributes to the 'backend' class
constructor. This was adding complexity (each backend would have to deal
with a unified set of constructor attributes), for little actual gain.
If one wants to create a really specific backend object, it's easier to
just create it and pass it directly to Git::Database->new.