Make no assumptions about the get_object_attributes return value
Up until now, it was implicitely assumed that the backend's
get_object_attributes() method always returned the `content`.

For tree, commit and tag objects, the content and the other data structure
(directory_entiers, commit_info, tag_info) can each be produced by the
other, so get_object_attributes could return either.

This patch rewrites the builders so that when an object is created
with only the digest, and the content or the data structure is
required, the attributes will be fetched via the backend (using
Git::Database::Role::Object's _get_object_attributes), and used to
populate the relevant object attributes (before continuing the build,
if necessary).