Enforce absolute directories when possible
When the store is built using `work_tree` or `git_dir`,
make sure that the backends that do not enfore an absolute
path internally will get one.

Note that this defensive measure is not taken when the
`store` parameter is given.

t/git-sub.t was renamed to t/relative.t because the test is
now much more about changing working directory after creating
the database object that just about the Git::Sub backend.