Detail the selection mechanism a bit more
The speed comparison was made using a quick script, that's probably not
very representative of actual performance.

It nevertheless provided results consistent with expectations, like
the fact that the fastest backend is backed by a C library, and the
following three only run one git process for fetching content from the
object database for the whole lifetime of the Git::Database object.